+++ src/Eloquent/Enumeration/Enumeration.php
@@ -11,62 +11,83 @@
 
 namespace Eloquent\Enumeration;
 
+use ReflectionClass;
+
 /**
- * Base class for C++ style enumerations.
+ * Abstract base class for C++ style enumerations.
  */
 abstract class Enumeration extends Multiton
 {
     /**
-     * Returns a single member instance by value.
+     * Returns a single member by value.
      *
-     * @param scalar $value The value associated with the member instance.
+     * @param scalar       $value           The value associated with the member.
+     * @param boolean|null $isCaseSensitive True if the search should be case sensitive.
      *
-     * @return Enumeration The first member instance with the supplied value.
-     * @throws Exception\UndefinedInstanceException If no associated instance is found.
+     * @return Enumeration                        The first member with the supplied value.
+     * @throws Exception\UndefinedMemberException If no associated member is found.
      */
-    public static final function instanceByValue($value)
+    final public static function memberByValue($value, $isCaseSensitive = null)
     {
-        foreach (static::multitonInstances() as $instance) {
-            if ($instance->value() === $value) {
-                return $instance;
-            }
-        }
-
-        throw static::createUndefinedInstanceException(get_called_class(), 'value', $value);
+        return static::memberBy('value', $value, $isCaseSensitive);
     }
 
     /**
-     * Returns the value of this member instance.
+     * Returns a single member by value. Additionally returns a default if no
+     * associated member is found.
      *
-     * @return scalar The value of this member instance.
+     * @param scalar        $value           The value associated with the member.
+     * @param Multiton|null $default         The default value to return.
+     * @param boolean|null  $isCaseSensitive True if the search should be case sensitive.
+     *
+     * @return Enumeration The first member with the supplied value, or the default value.
      */
-    public final function value()
+    final public static function memberByValueWithDefault(
+        $value,
+        Multiton $default = null,
+        $isCaseSensitive = null
+    ) {
+        return static::memberByWithDefault(
+            'value',
+            $value,
+            $default,
+            $isCaseSensitive
+        );
+    }
+
+    /**
+     * Returns the value of this member.
+     *
+     * @return scalar The value of this member.
+     */
+    final public function value()
     {
         return $this->value;
     }
 
     /**
-     * Initializes the member instances of this enumeration based upon its class
+     * Initializes the members of this enumeration based upon its class
      * constants.
      *
-     * Each constant becomes a member instance with a string key equal to the
-     * constant's name, and a value equal to that of the constant's value.
+     * Each constant becomes a member with a string key equal to the constant's
+     * name, and a value equal to that of the constant's value.
      */
-    protected static final function initializeMultiton()
+    final protected static function initializeMembers()
     {
-        $reflector = new \ReflectionClass(get_called_class());
+        $reflector = new ReflectionClass(get_called_class());
+
         foreach ($reflector->getConstants() as $key => $value) {
             new static($key, $value);
         }
     }
 
     /**
-     * Construct and register a new enumeration member instance.
+     * Construct and register a new enumeration member.
      *
-     * @param string $key The string key to associate with this member instance.
-     * @param scalar $value The value of this member instance.
+     * @param string $key   The string key to associate with this member.
+     * @param scalar $value The value of this member.
      *
-     * @throws Exception\ExtendsConcreteException If the constructed instance has an invalid inheritance hierarchy.
+     * @throws Exception\ExtendsConcreteException If the constructed member has an invalid inheritance hierarchy.
      */
     protected function __construct($key, $value)
     {
@@ -76,7 +97,7 @@ abstract class Enumeration extends Multiton
     }
 
     /**
-     * The value of this member instance.
+     * The value of this member.
      *
      * @var scalar
      */