+++ src/Eloquent/Enumeration/Multiton.php
@@ -11,138 +11,237 @@
 
 namespace Eloquent\Enumeration;
 
-use Eloquent\Enumeration\Exception\ExtendsConcreteException;
-use Eloquent\Enumeration\Exception\UndefinedInstanceException;
-use Exception;
+use Exception as NativeException;
+use ReflectionObject;
 
 /**
- * Base class for Java style enumerations.
+ * Base class for Java-style enumerations.
  */
 abstract class Multiton
 {
     /**
-     * Returns an array of all member instancess in this multiton.
+     * Returns an array of all members in this multiton.
      *
-     * @return array<string,Multiton> All member instancess in this multiton.
+     * @return array<string,Multiton> All members.
      */
-    public static final function multitonInstances()
+    final public static function members()
     {
         $class = get_called_class();
-        if (!array_key_exists($class, self::$instances)) {
-            self::$instances[$class] = array();
-            static::initializeMultiton();
+        if (!array_key_exists($class, self::$members)) {
+            self::$members[$class] = array();
+            static::initializeMembers();
         }
 
-        return self::$instances[$class];
+        return self::$members[$class];
     }
 
     /**
-     * Returns a single member instance by string key.
+     * Returns a single member by string key.
      *
-     * @param string $key The string key associated with the member instance.
+     * @param string       $key             The string key associated with the member.
+     * @param boolean|null $isCaseSensitive True if the search should be case sensitive.
      *
-     * @return Multiton The member instance associated with the given string key.
-     * @throws UndefinedInstanceException If no associated instance is found.
+     * @return Multiton                                    The member associated with the given string key.
+     * @throws Exception\UndefinedMemberExceptionInterface If no associated member is found.
      */
-    public static final function instanceByKey($key)
+    final public static function memberByKey($key, $isCaseSensitive = null)
     {
-        $instances = static::multitonInstances();
-        if (array_key_exists($key, $instances)) {
-            return $instances[$key];
-        }
-
-        throw static::createUndefinedInstanceException(get_called_class(), 'key', $key);
+        return static::memberBy('key', $key, $isCaseSensitive);
     }
 
     /**
-     * Returns a single member instance by comparison with the result of an accessor method.
+     * Returns a single member by string key. Additionally returns a default if
+     * no associated member is found.
      *
-     * @param string $property The name of the property (accessor method) to match.
-     * @param mixed $value The value to match.
+     * @param string        $key             The string key associated with the member.
+     * @param Multiton|null $default         The default value to return.
+     * @param boolean|null  $isCaseSensitive True if the search should be case sensitive.
      *
-     * @return Multiton The first member instance for which $instance->{$property}() === $value.
-     * @throws UndefinedInstanceException If no associated instance is found.
+     * @return Multiton The member associated with the given string key, or the default value.
      */
-    public static final function instanceBy($property, $value) {
-        foreach (static::multitonInstances() as $instance) {
-            if ($instance->{$property}() === $value) {
-                return $instance;
+    final public static function memberByKeyWithDefault(
+        $key,
+        Multiton $default = null,
+        $isCaseSensitive = null
+    ) {
+        return static::memberByWithDefault(
+            'key',
+            $key,
+            $default,
+            $isCaseSensitive
+        );
+    }
+
+    /**
+     * Returns a single member by comparison with the result of an accessor
+     * method.
+     *
+     * @param string       $property        The name of the property (accessor method) to match.
+     * @param mixed        $value           The value to match.
+     * @param boolean|null $isCaseSensitive True if the search should be case sensitive.
+     *
+     * @return Multiton                                    The first member for which $member->{$property}() === $value.
+     * @throws Exception\UndefinedMemberExceptionInterface If no associated member is found.
+     */
+    final public static function memberBy(
+        $property,
+        $value,
+        $isCaseSensitive = null
+    ) {
+        $member = static::memberByWithDefault(
+            $property,
+            $value,
+            null,
+            $isCaseSensitive
+        );
+        if (null === $member) {
+            throw static::createUndefinedMemberException(
+                get_called_class(),
+                $property,
+                $value
+            );
+        }
+
+        return $member;
+    }
+
+    /**
+     * Returns a single member by comparison with the result of an accessor
+     * method. Additionally returns a default if no associated member is found.
+     *
+     * @param string        $property        The name of the property (accessor method) to match.
+     * @param mixed         $value           The value to match.
+     * @param Multiton|null $default         The default value to return.
+     * @param boolean|null  $isCaseSensitive True if the search should be case sensitive.
+     *
+     * @return Multiton|null The first member for which $member->{$property}() === $value, or the default value.
+     */
+    final public static function memberByWithDefault(
+        $property,
+        $value,
+        Multiton $default = null,
+        $isCaseSensitive = null
+    ) {
+        if (null === $isCaseSensitive) {
+            $isCaseSensitive = true;
+        }
+        if (!$isCaseSensitive && is_scalar($value)) {
+            $value = strtoupper(strval($value));
+        }
+
+        return static::memberByPredicateWithDefault(
+            function (Multiton $member) use (
+                $property,
+                $value,
+                $isCaseSensitive
+            ) {
+                $memberValue = $member->{$property}();
+                if (!$isCaseSensitive && is_scalar($memberValue)) {
+                    $memberValue = strtoupper(strval($memberValue));
+                }
+
+                return $memberValue === $value;
+            },
+            $default
+        );
+    }
+
+    /**
+     * Returns a single member by predicate callback.
+     *
+     * @param callback $predicate The predicate applies to the member to find a match.
+     *
+     * @return Multiton                                    The first member for which $predicate($member) evaluates to boolean true.
+     * @throws Exception\UndefinedMemberExceptionInterface If no associated member is found.
+     */
+    final public static function memberByPredicate($predicate)
+    {
+        $member = static::memberByPredicateWithDefault($predicate);
+        if (null === $member) {
+            throw static::createUndefinedMemberException(
+                get_called_class(),
+                '<callback>',
+                '<callback>'
+            );
+        }
+
+        return $member;
+    }
+
+    /**
+     * Returns a single member by predicate callback. Additionally returns a
+     * default if no associated member is found.
+     *
+     * @param callback      $predicate The predicate applies to the member to find a match.
+     * @param Multiton|null $default   The default value to return.
+     *
+     * @return Multiton The first member for which $predicate($member) evaluates to boolean true, or the default value.
+     */
+    final public static function memberByPredicateWithDefault(
+        $predicate,
+        Multiton $default = null
+    ) {
+        foreach (static::members() as $member) {
+            if ($predicate($member)) {
+                return $member;
             }
         }
 
-        throw static::createUndefinedInstanceException(get_called_class(), $property, $value);
+        return $default;
     }
 
     /**
-     * Returns a single member instance by predicate callback.
+     * Maps static method calls to members.
      *
-     * @param callback $predicate The predicate applies to the multiton instance to find a match.
+     * @param string $key       The string key associated with the member.
+     * @param array  $arguments Ignored.
      *
-     * @return Multiton The first member instance for which $predicate($instance) evaluates to boolean true.
-     * @throws UndefinedInstanceException If no associated instance is found.
+     * @return Multiton                                    The member associated with the given string key.
+     * @throws Exception\UndefinedMemberExceptionInterface If no associated member is found.
      */
-    public static final function instanceByPredicate($predicate) {
-        foreach (static::multitonInstances() as $instance) {
-            if ($predicate($instance)) {
-                return $instance;
-            }
-        }
-
-        throw static::createUndefinedInstanceException(get_called_class(), '<callback>', '<callback>');
-    }
-
-    /**
-     * Maps static method calls to member instances.
-     *
-     * @param string $key The string key associated with the member instance.
-     * @param array $arguments Ignored.
-     *
-     * @return Multiton The member instance associated with the given string key.
-     * @throws UndefinedInstanceException If no associated instance is found.
-     */
-    public static final function __callStatic($key, array $arguments)
+    final public static function __callStatic($key, array $arguments)
     {
-        return static::instanceByKey($key);
+        return static::memberByKey($key);
     }
 
     /**
-     * Returns the string key of this member instance.
+     * Returns the string key of this member.
      *
-     * @return string The associated string key of this member instance.
+     * @return string The associated string key of this member.
      */
-    public final function key()
+    final public function key()
     {
         return $this->key;
     }
 
     /**
-     * Check if this multiton instance is in the specified list of values.
+     * Check if this member is in the specified list of members.
      *
-     * @param Multiton $a     The first multiton instance to check.
-     * @param Multiton $b     The second multiton instance to check.
-     * @param Multiton $c,... Additional multiton instances to check.
+     * @param Multiton $a     The first member to check.
+     * @param Multiton $b     The second member to check.
+     * @param Multiton $c,... Additional members to check.
      *
-     * @return boolean True if this multiton instance value is in the specified list of values.
+     * @return boolean True if this member is in the specified list of members.
      */
-    public final function anyOf(Multiton $a, Multiton $b)
+    final public function anyOf(Multiton $a, Multiton $b)
     {
         return $this->anyOfArray(func_get_args());
     }
 
     /**
-     * Check if this multiton instance is in the specified list of values.
+     * Check if this member is in the specified list of members.
      *
-     * @param array<Multiton> $values An array of multion values to search.
+     * @param array<Multiton> $values An array of members to search.
      *
-     * @return boolean True if this multiton instance value is in the specified list of values.
+     * @return boolean True if this member is in the specified list of members.
      */
-    public final function anyOfArray(array $values)
+    final public function anyOfArray(array $values)
     {
         return in_array($this, $values, true);
     }
 
     /**
-     * Returns a string representation of this member instance.
+     * Returns a string representation of this member.
      *
      * Unless overridden, this is simply the string key.
      *
@@ -154,89 +253,94 @@ abstract class Multiton
     }
 
     /**
-     * Override this method in child classes to implement one-time initialization
-     * for a multiton class.
+     * Override this method in child classes to implement one-time
+     * initialization for a multiton class.
      *
      * This method is called the first time the members of a multiton are
      * accessed. It is called via late static binding, and hence can be
      * overridden in child classes.
      */
-    protected static function initializeMultiton() {}
+    protected static function initializeMembers() {}
 
     /**
      * Override this method in child classes to implement custom undefined
-     * instance exceptions for a multiton class.
+     * member exceptions for a multiton class.
      *
-     * @param string $className
-     * @param string $property
-     * @param mixed $value
-     * @param Exception|null $previous
+     * @param string               $className
+     * @param string               $property
+     * @param mixed                $value
+     * @param NativeException|null $previous
      *
-     * @return UndefinedInstanceExceptionInterface
+     * @return UndefinedMemberExceptionInterface
      */
-    protected static function createUndefinedInstanceException(
+    protected static function createUndefinedMemberException(
         $className,
         $property,
         $value,
-        Exception $previous = null
+        NativeException $previous = null
     ) {
-        return new UndefinedInstanceException($className, $property, $value, $previous);
+        return new Exception\UndefinedMemberException(
+            $className,
+            $property,
+            $value,
+            $previous
+        );
     }
 
     /**
-     * Construct and register a new multiton member instance.
+     * Construct and register a new multiton member.
      *
      * If you override the constructor in a child class, you MUST call the parent
      * constructor. Calling this constructor is the only way to set the string
-     * key for this member instance, and to ensure that the instance is correctly
+     * key for this member, and to ensure that the member is correctly
      * registered.
      *
-     * @param string $key The string key to associate with this member instance.
+     * @param string $key The string key to associate with this member.
      *
-     * @throws ExtendsConcreteException If the constructed instance has an invalid inheritance hierarchy.
+     * @throws Exception\ExtendsConcreteException If the constructed member has an invalid inheritance hierarchy.
      */
     protected function __construct($key)
     {
         $this->key = $key;
 
-        self::registerMultiton($this);
+        self::registerMember($this);
     }
 
     /**
-     * Registers the supplied member instance.
+     * Registers the supplied member.
      *
      * Do not attempt to call this method directly. Instead, ensure that
      * Multiton::__construct() is called from any child classes, as this will
-     * also handle registration of the instance.
+     * also handle registration of the member.
      *
-     * @param Multiton $instance The instance to register.
-     * @throws ExtendsConcreteException If the supplied instance has an invalid inheritance hierarchy.
+     * @param  Multiton                           $member The member to register.
+     * @throws Exception\ExtendsConcreteException If the supplied member has an invalid inheritance hierarchy.
      */
-    private static function registerMultiton(self $instance)
+    private static function registerMember(self $member)
     {
-        $reflector = new \ReflectionObject($instance);
+        $reflector = new ReflectionObject($member);
         $parentClass = $reflector->getParentClass();
         if (!$parentClass->isAbstract()) {
-            throw new ExtendsConcreteException(
-                get_class($instance)
-                , $parentClass->getName()
+            throw new Exception\ExtendsConcreteException(
+                get_class($member),
+                $parentClass->getName()
             );
         }
 
-        self::$instances[get_called_class()][$instance->key()] = $instance;
+        self::$members[get_called_class()][$member->key()] = $member;
     }
 
     /**
-     * Array of all member instances of all multiton and enumeration classes.
+     * Array of all members of all multiton and enumeration classes.
      *
-     * Instances are keyed by class name and member key string.
+     * Members are keyed by class name and member key string.
      *
      * @var array<string,array<string,Multiton>>
      */
-    private static $instances = array();
+    private static $members = array();
 
     /**
-     * String key associated with this member instance.
+     * String key associated with this member.
      *
      * @var string
      */