+++ test/bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-
-// include fixtures than cannot be autoloaded
-$documentationFixturePath =
-    __DIR__.'/src/Eloquent/Enumeration/Test/Fixture/Documentation'
-;
-require $documentationFixturePath.'/HTTPRequestMethod.php';
-require $documentationFixturePath.'/Planet.php';